добавлен HOC withVideoPlayer
