+++ src/components/video-player/video-player.jsx
@@ -1,92 +1,22 @@
-import React, {useRef, useEffect} from "react";
+import React from "react";
 import PropTypes from "prop-types";
-import {FilmCardShape} from "../../constants";
 
-export default class VideoPlayer extends React.PureComponent {
-  constructor(props) {
-    super(props);
-    this._videoRef = React.createRef();
-    this.state = {
-      progress: 0,
-      volume: 0,
-      isLoading: true,
-      isPlaying: this.props.isPlaying,
-    };
-  }
-
-  static getDerivedStateFromProps(props, state) {
-    if (props.isPlaying !== state.isPlaying) {
-      return {isPlaying: props.isPlaying};
-    }
-    return null;
-  }
-
-  componentDidMount() {
-    const {src, poster} = this.props;
-    const video = this._videoRef.current;
-    video.src = src;
-    video.poster = poster;
-
-    video.oncanplaythrough = () => this.setState({
-      isLoading: false,
-    });
-
-    video.onplay = () => this.setState({
-      isPlaying: true,
-    });
-
-    video.onpause = () => this.setState({
-      isPlaying: false,
-    });
-
-    video.ontimeupdate = () => this.setState({
-      progress: video.currentTime,
-    });
-
-    video.onvolumechange = () => this.setState({
-      volume: video.volume,
-    });
-  }
-
-  componentWillUnmount() {
-    const video = this._videoRef.current;
-
-    video.oncanplaythrough = null;
-    video.onplay = null;
-    video.onpause = null;
-    video.ontimeupdate = null;
-    video.onvolumechange = null;
-    video.src = ``;
-    video.poster = ``;
-  }
-
+class VideoPlayer extends React.PureComponent {
   render() {
-    const {src, poster} = this.props;
-    const isMuted = this.state.volume === 0;
+    const {children} = this.props;
     return (
-      <video
-        ref={this._videoRef}
-        src={src}
-        poster={poster}
-        muted={isMuted}
-        width={FilmCardShape.WIDTH}
-        height={FilmCardShape.HEIGHT}
-      />
+      <>
+        {children}
+      </>
     );
-  }
-
-  componentDidUpdate() {
-    const video = this._videoRef.current;
-    if (this.state.isPlaying) {
-      video.play();
-    } else {
-      video.load();
-    }
   }
 }
 
 VideoPlayer.propTypes = {
-  isPlaying: PropTypes.bool.isRequired,
-  src: PropTypes.string.isRequired,
-  poster: PropTypes.string.isRequired,
+  children: PropTypes.oneOfType([
+    PropTypes.arrayOf(PropTypes.node),
+    PropTypes.node
+  ]).isRequired,
 };
+
+export default VideoPlayer;
